fix(atelier): keep deep template nesting off the machine stack
`MAX_ELEMENT_NESTING_DEPTH` was 256 and could not simply be raised. The
transform lane, static hoisting, codegen, slot/namespace detection, the
patterned-template and legacy pre-passes, the parser's whitespace condenser
and the AST's own teardown each spend one machine-stack frame per nesting
level, so one level past the constant the failure mode was `fatal runtime
error: stack overflow` — SIGABRT, which takes the whole process down and
cannot be turned into a diagnostic. 256 was simply the depth a debug build
survived on Rust's default 2 MiB thread stack, which left Vize below
`@vue/compiler-dom`: upstream reaches ~1092 levels of `<div>` on a default
Node stack and then raises a *catchable* `RangeError`.

Add `vize_carton::recursion::ensure_sufficient_stack`, a thin
`stacker::maybe_grow` wrapper, and wrap every descent whose depth is bounded
by the input rather than by the code. Nesting depth now costs heap, which is
recoverable and bounded by the parser's limit, instead of stack, which is
not. An explicit worklist is the cleaner shape and stays the direction for
the transform lane, but codegen interleaves emission with recursion at ~20
call sites across ten modules, so making the whole middle end iterative is a
continuation-passing rewrite; guarding the recursion covers all of it with
one mechanism and no behavioural risk, and is the same answer rustc reached
(`rustc_data_structures::stack::ensure_sufficient_stack`).

With the stack no longer binding, raise the limit to 4096 — ~3.7x upstream's
measured ceiling — chosen for what it now actually bounds, output size:
generated code is quadratic in nesting depth, so 4096 is already ~34 MB.
Exceeding it stays a recoverable diagnostic reported once per over-limit
region.

Splits `hoist_static.rs`, `codegen/generate.rs` and `transforms/v_slot.rs`
into submodules so no over-limit source file grows.

Closes #3480
