+++ crates/vize_carton/Cargo.toml
@@ -21,6 +21,7 @@ serde = { workspace = true }
 serde_json = { workspace = true }
 once_cell = { workspace = true }
 smallvec = { workspace = true }
+stacker = { workspace = true }
 xxhash-rust = { workspace = true }
 pklrust = { workspace = true }
